fix type Task1

+++ tyuiu.cources.programming.interfaces/ISprint3Task1.cs
@@ -15,23 +15,23 @@ namespace tyuiu.cources.programming.interfaces.Sprint3
     public interface ISprint3Task1V6 { double GetMultiplySeries(); }
     public interface ISprint3Task1V7 { double GetMultiplySeries(double value); }
     public interface ISprint3Task1V8 { double GetMultiplySeries(double value); }
-    public interface ISprint3Task1V9 { double GetMultiplySeries(int value); }
-    public interface ISprint3Task1V10 { double GetMultiplySeries(int value); }
+    public interface ISprint3Task1V9 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V10 { double GetMultiplySeries(double value); }
     public interface ISprint3Task1V11 { double GetMultiplySeries(double value); }
-    public interface ISprint3Task1V12 { double GetMultiplySeries(int value); }
+    public interface ISprint3Task1V12 { double GetMultiplySeries(double value); }
     public interface ISprint3Task1V13 { double GetMultiplySeries(double value); }
-    public interface ISprint3Task1V14 { int GetMultiplySeries(int value); }
-    public interface ISprint3Task1V15 { int GetMultiplySeries(int value); }
-    public interface ISprint3Task1V16 { int GetMultiplySeries(); }
-    public interface ISprint3Task1V17 { double GetMultiplySeries(int value); }
-    public interface ISprint3Task1V18 { double GetMultiplySeries(int value); }
-    public interface ISprint3Task1V19 { double GetMultiplySeries(int value); }
-    public interface ISprint3Task1V20 { double GetMultiplySeries(int value); }
-    public interface ISprint3Task1V21 { double GetMultiplySeries(int value); }
+    public interface ISprint3Task1V14 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V15 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V16 { double GetMultiplySeries(); }
+    public interface ISprint3Task1V17 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V18 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V19 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V20 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V21 { double GetMultiplySeries(double value); }
     public interface ISprint3Task1V22 { double GetMultiplySeries(double value); }
-    public interface ISprint3Task1V23 { double GetMultiplySeries(int value); }
-    public interface ISprint3Task1V24 { double GetMultiplySeries(int value); }
-    public interface ISprint3Task1V25 { double GetMultiplySeries(int value); }
+    public interface ISprint3Task1V23 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V24 { double GetMultiplySeries(double value); }
+    public interface ISprint3Task1V25 { double GetMultiplySeries(double value); }
     public interface ISprint3Task1V26 { double GetMultiplySeries(); }
     public interface ISprint3Task1V27 { double GetMultiplySeries(double value); }
     public interface ISprint3Task1V28 { double GetMultiplySeries(double value); }
